+++ admin.php
@@ -1,8 +1,8 @@
 <?php
 /**
- *  Admin Extensions for Leaflet Map DSGVO
+ *  Admin DSGVO for Leaflet Map and its Extensions
  *
- * @package Extensions for Leaflet Map DSGVO
+ * @package DSGVO snippet for Leaflet Map and its Extensions
  **/
 
 // Direktzugriff auf diese Datei verhindern.
@@ -13,8 +13,8 @@ function leafext_dsgvo_add_page() {
 	// Add Submenu
 	$leafext_admin_page = add_submenu_page(
 		'leaflet-map',
-		__( 'Leaflet Map Options GDPR', 'extensions-leaflet-map-dsgvo' ),
-		__( 'Leaflet Map GDPR', 'extensions-leaflet-map-dsgvo' ),
+		'Leaflet Map ' . __( 'Options GDPR', 'dsgvo-leaflet-map' ),
+		'Leaflet Map ' . __( 'GDPR', 'dsgvo-leaflet-map' ),
 		'manage_options',
 		LEAFEXT_DSGVO_PLUGIN_NAME,
 		'leafext_dsgvo_do_page'
@@ -45,12 +45,10 @@ function leafext_dsgvo_init() {
 add_action( 'admin_init', 'leafext_dsgvo_init' );
 
 function leafext_dsgvo_form( $field ) {
-	// var_dump($field);
 	$options  = leafext_dsgvo_params();
 	$option   = leafext_array_find3( $field, $options );
 	$settings = leafext_dsgvo_settings();
-	// var_dump($settings,$option);
-	$setting = $settings[ $field ];
+	$setting  = $settings[ $field ];
 	if ( is_plugin_active( 'theme-translation-for-polylang/polylang-theme-translation.php' ) ) {
 		$ttfp = ' readonly ';
 	} else {
@@ -58,37 +56,34 @@ function leafext_dsgvo_form( $field ) {
 	}
 	switch ( $field ) {
 		case 'text':
-			//phpcs:ignore WordPress.Security.EscapeOutput.OutputNotEscaped -- string not changeable
-			echo '<textarea ' . $ttfp . ' name="leafext_dsgvo[text]" type="textarea" cols="80" rows="5">';
-			echo esc_textarea( $setting );
+			echo '<textarea ' . esc_attr( $ttfp ) . ' name="leafext_dsgvo[text]" type="textarea" cols="80" rows="5">';
+			echo wp_kses_post( $setting );
 			echo '</textarea>';
 			break;
 		case 'mapurl':
-			echo '<input type="url" size="80" name="leafext_dsgvo[mapurl]" value="' . esc_url( $setting ) .
-			'" />';
+			echo '<input type="url" size="80" name="leafext_dsgvo[mapurl]" value="' . esc_url( $setting ) . '" />';
 			break;
 		case 'color':
 			leafext_dsgvo_colors( $option['default'], $setting );
 			break;
 		case 'cookie':
-			echo '<input type="number" size="5" min="1" max="365" name="leafext_dsgvo[cookie]" value=' . esc_textarea( $setting ) . '> ';
-			esc_html_e( 'days', 'extensions-leaflet-map-dsgvo' );
+			echo '<input type="number" size="5" min="1" max="365" name="leafext_dsgvo[cookie]" value="' . absint( $setting ) . '"> ';
+			esc_html_e( 'days', 'dsgvo-leaflet-map' );
 			break;
 		case 'count':
 			echo '<input type="radio" name="leafext_dsgvo[count]" value="1" ';
 			echo boolval( $setting ) ? 'checked' : '';
 			echo '> ';
-			esc_html_e( 'each map', 'extensions-leaflet-map-dsgvo' );
+			esc_html_e( 'each map', 'dsgvo-leaflet-map' );
 			echo ' &nbsp;&nbsp; ';
 			echo '<input type="radio" name="leafext_dsgvo[count]" value="0" ';
 			echo ! boolval( $setting ) ? 'checked' : '';
 			echo '> ';
-			esc_html_e( 'only first', 'extensions-leaflet-map-dsgvo' );
+			esc_html_e( 'only first', 'dsgvo-leaflet-map' );
 			echo ' ';
 			break;
 		case 'okay':
-			//phpcs:ignore WordPress.Security.EscapeOutput.OutputNotEscaped -- string not changeable
-			echo '<input type="text" ' . $ttfp . ' size="10" name="leafext_dsgvo[okay]" value="' . esc_textarea( $setting ) . '" />';
+			echo '<input type="text" ' . esc_attr( $ttfp ) . ' size="10" name="leafext_dsgvo[okay]" value="' . esc_attr( $setting ) . '" />';
 			break;
 		default:
 			wp_die( 'error' );
@@ -105,22 +100,22 @@ function leafext_validate_dsgvo( $options ) {
 			$defaults[ $param['param'] ] = $param['default'];
 		}
 		if ( isset( $options['cookie'] ) && ( $options['cookie'] === '0' || $options['cookie'] === '' ) ) {
-			$options['cookie'] = $defaults['cookie'];
+			$options['cookie'] = absint( $defaults['cookie'] );
 		}
 		if ( isset( $options['text'] ) ) {
-			$options['text'] = wp_kses_normalize_entities( $options['text'] );
+			$options['text'] = wp_kses_post( $options['text'] );
 		}
 		if ( isset( $options['mapurl'] ) ) {
-			$options['mapurl'] = sanitize_text_field( $options['mapurl'] );
+			$options['mapurl'] = sanitize_url( $options['mapurl'] );
 		}
 		if ( isset( $options['color'] ) ) {
 			$options['color'] = sanitize_text_field( $options['color'] );
 		}
 		if ( isset( $options['count'] ) ) {
-			$options['count'] = $options['count'];
+			$options['count'] = absint( $options['count'] );
 		}
 		if ( isset( $options['okay'] ) ) {
-			$options['okay'] = wp_kses_normalize_entities( $options['okay'] );
+			$options['okay'] = sanitize_text_field( $options['okay'] );
 		}
 		$change = array();
 		foreach ( $options as $key => $value ) {
@@ -128,7 +123,6 @@ function leafext_validate_dsgvo( $options ) {
 				$change[ $key ] = $value;
 			}
 		}
-		//var_dump($options,$defaults,$change); wp_die();
 		return $change;
 	}
 	if ( isset( $_POST['delete'] ) ) {
@@ -139,72 +133,42 @@ function leafext_validate_dsgvo( $options ) {
 
 // Erklaerung / Hilfe
 function leafext_dsgvo_help() {
-	// Call globals
-	global $wp_filesystem;
-	// Initiate
-	WP_Filesystem();
-	$local_file = LEAFEXT_DSGVO_PLUGIN_DIR . '/readme.md';
-	$text       = '';
-	if ( $wp_filesystem->exists( $local_file ) ) {
-		$text = $wp_filesystem->get_contents( $local_file );
-
-		$suchmuster = array(
-			'/(.+)## Description/s',
-			'/\[(.+)\]\((.+)\)/i',
-			'/(### )(.*)/',
-			'/(## )(.*)/',
-			'/(# )(.*)/',
-			'/  /',
-		);
-		$ersetzung  = array(
-			'## ' . __( 'Description', 'extensions-leaflet-map-dsgvo' ),
-			'<a href="${2}">${1}</a>',
-			'<h3>${2}</h3>',
-			'<h2>${2}</h2>',
-			'<h1>${2}</h1>',
-			'<br>',
-		);
-		$text       = preg_replace( $suchmuster, $ersetzung, $text );
-		// https://wp-mix.com/allowed-html-tags-wp_kses/
-		$allowed_tags = wp_kses_allowed_html( 'post' );
-		echo '<div style="width:80%">' . wp_kses( $text, $allowed_tags ) . '</div>';
-	} else {
-		echo 'Error';
-	}
+	$text = '<h3>' . sprintf(
+		/* translators: %s is Leaflet Map */
+		__( 'GDPR (DSGVO) snippet for %s and its Extensions', 'dsgvo-leaflet-map' ),
+		'Leaflet Map'
+	) . '</h3>';
+	echo wp_kses_post( $text );
 }
 
 function leafext_dsgvo_ttfp_help() {
-	if ( is_plugin_active( 'polylang/polylang.php' ) && is_plugin_active( 'leaflet-map/leaflet-map.php' ) ) {
+	if ( defined( 'POLYLANG_VERSION' ) ) {
 		echo '<h3>Polylang</h3>';
-		$ttfp = '<a href="https://wordpress.org/plugins/theme-translation-for-polylang/">Theme and plugin translation for Polylang (TTfP)</a> ';
+		$ttfp = '<a href="' . esc_url( 'https://wordpress.org/plugins/theme-translation-for-polylang/' ) . '">Theme and plugin translation for Polylang (TTfP)</a> ';
 		if ( is_plugin_active( 'theme-translation-for-polylang/polylang-theme-translation.php' ) ) {
 			echo '<ul><li>';
-			//phpcs:ignore WordPress.Security.EscapeOutput.OutputNotEscaped -- string not changeable
-			echo $ttfp . ' ';
+			echo wp_kses_post( $ttfp ) . ' ';
 			$ttfp = true;
-			esc_html_e( 'is active.', 'extensions-leaflet-map-dsgvo' );
+			esc_html_e( 'is active.', 'dsgvo-leaflet-map' );
 			echo '</li><li>';
-			esc_html_e( 'Go to', 'extensions-leaflet-map-dsgvo' );
-			//phpcs:ignore WordPress.Security.EscapeOutput.OutputNotEscaped -- is a function
-			echo ' <a href="' . admin_url( 'admin.php' ) . '?page=mlang_import_export_strings">';
-			esc_html_e( 'Settings', 'extensions-leaflet-map-dsgvo' );
+			esc_html_e( 'Go to', 'dsgvo-leaflet-map' );
+			echo ' <a href="' . esc_url( admin_url( 'admin.php' ) . '?page=mlang_import_export_strings' ) . '">';
+			esc_html_e( 'Settings', 'dsgvo-leaflet-map' );
 			echo '</a>, ';
-			esc_html_e( 'enable', 'extensions-leaflet-map-dsgvo' );
-			echo ' <code>leafext-dsgvo</code> ';
-			esc_html_e( 'and', 'extensions-leaflet-map-dsgvo' );
-			//phpcs:ignore WordPress.Security.EscapeOutput.OutputNotEscaped -- is a function
-			echo ' <a href="' . admin_url( 'admin.php' ) . '?page=mlang_strings&s&group=TTfP%3A+leafext-dsgvo&paged=1">';
-			esc_html_e( 'fill in your text', 'extensions-leaflet-map-dsgvo' );
+			esc_html_e( 'enable', 'dsgvo-leaflet-map' );
+			echo ' <code>' . esc_html( LEAFEXT_DSGVO_PLUGIN_NAME ) . '</code> ';
+			esc_html_e( 'and', 'dsgvo-leaflet-map' );
+			echo ' <a href="' . esc_url( admin_url( 'admin.php' ) . '?page=mlang_strings&s&group=TTfP%3A+' . LEAFEXT_DSGVO_PLUGIN_NAME ) . '">';
+			esc_html_e( 'fill in your text', 'dsgvo-leaflet-map' );
 			echo '</a>!';
 			echo '</li></ul>';
 		} else {
 			printf(
 				/* translators: %s is a link. */
-				esc_html__( 'If you wish to translate these strings in %s use', 'extensions-leaflet-map-dsgvo' ),
-				' <a href="https://wordpress.org/plugins/polylang/">Polylang</a> '
+				esc_html__( 'If you wish to translate these strings in %s use', 'dsgvo-leaflet-map' ),
+				' <a href="' . esc_url( 'https://wordpress.org/plugins/polylang/' ) . '">Polylang</a> '
 			);
-			//phpcs:ignore WordPress.Security.EscapeOutput.OutputNotEscaped -- string not changeable
-			echo ' ' . $ttfp . '.';
+			echo ' ' . wp_kses_post( $ttfp ) . '.';
 			$ttfp = false;
 		}
 	}
@@ -214,14 +178,14 @@ function leafext_dsgvo_ttfp_help() {
 function leafext_dsgvo_do_page() {
 	leafext_dsgvo_help();
 	leafext_dsgvo_update_admin();
-		echo '<h3>';
-	esc_html_e( 'Settings', 'extensions-leaflet-map-dsgvo' );
+	echo '<h3>';
+	esc_html_e( 'Settings', 'dsgvo-leaflet-map' );
 	echo '</h3>';
 	echo '<p>';
 	if ( is_plugin_active( 'leaflet-map/leaflet-map.php' ) ) {
-		esc_html_e( 'Test it in a private browser window.', 'extensions-leaflet-map-dsgvo' );
+		esc_html_e( 'Test it in a private browser window.', 'dsgvo-leaflet-map' );
 	} else {
-		esc_html_e( 'Leaflet Map is not active.', 'extensions-leaflet-map-dsgvo' );
+		esc_html_e( 'Leaflet Map is not active.', 'dsgvo-leaflet-map-github' );
 	}
 	echo '</p>';
 	if ( is_plugin_active( 'leaflet-map/leaflet-map.php' ) ) {
@@ -232,7 +196,7 @@ function leafext_dsgvo_do_page() {
 		do_settings_sections( 'leafext_settings_dsgvo' );
 		if ( current_user_can( 'manage_options' ) ) {
 			submit_button();
-			submit_button( __( 'Reset', 'extensions-leaflet-map-dsgvo' ), 'delete', 'delete', false );
+			submit_button( __( 'Reset', 'dsgvo-leaflet-map' ), 'delete', 'delete', false );
 		}
 		echo '</form>';
 	}
@@ -255,20 +219,17 @@ function leafext_dsgvo_colors( $defcolor, $value ) {
 		'wp-color-picker-alpha',
 		plugins_url( '/js/wp-color-picker-alpha.min.js', __FILE__ ),
 		array( 'wp-color-picker' ),
-		null,
+		LEAFEXT_DSGVO_PLUGIN_VERSION,
 		true
 	);
 	wp_enqueue_script(
 		'leafext-picker',
 		plugins_url( '/js/colorpicker.js', __FILE__ ),
 		array( 'wp-color-picker-alpha', 'wp-color-picker' ),
-		null,
+		LEAFEXT_DSGVO_PLUGIN_VERSION,
 		true
 	);
 
-	// phpcs:disable WordPress.Security.EscapeOutput.OutputNotEscaped
-		echo '<input type="text" class="color-picker" id="leafext_dsgvo_color" name="leafext_dsgvo[color]"
-			data-alpha-enabled="true" data-default-color="' . $defcolor . '" value="' . $value . '">';
-
-	// phpcs:enable WordPress.Security.EscapeOutput.OutputNotEscaped
+	echo '<input type="text" class="color-picker" id="leafext_dsgvo_color" name="leafext_dsgvo[color]" data-alpha-enabled="true" data-default-color="'
+	. esc_attr( $defcolor ) . '" value="' . esc_attr( $value ) . '">';
 }